# optimize CBOR string emitter