@f2prateek 2017-05-18
I'm definitely missing something here. Why not just do:

```
io.WriteString(e.w, v)
```