@achille-roussel 2017-05-18
The implementation of io.WriteString is
```go
func WriteString(w Writer, s string) (int, error) {
  if sw, ok := w.(stringWriter); ok {
    return sw.WriteString(s)
  }
  return w.Write([]byte(s))
}
```
So if the the writer doesn't implement the `stringWriter` interface there's a copy and a memory allocation being made when converting the string to a byte slice.

The change suggested in the PR will never cause a dynamic memory allocation nor a copy.